Plugin-helper: add more funcs
- Adds Done, WriteDone, and GetResultsDir functions
seperate from the results writer itself.
