+++ plugin-helper/helper.go
@@ -1,73 +1,37 @@
 package plugin_helper
 
 import (
-	"github.com/pkg/errors"
-	"gopkg.in/yaml.v2"
+	"fmt"
 	"os"
 	"path/filepath"
 
-	sono "github.com/vmware-tanzu/sonobuoy/pkg/client/results"
+	"github.com/vmware-tanzu/sonobuoy/pkg/tarball"
 )
 
 const(
-	defaultOutputFileName = "manual_results.yaml"
+	SonobuoyResultsDirKey="SONOBUOY_RESULTS_DIR"
+	DoneFileName = "done"
+	DefaultTarballName = "results.tar.gz"
 )
 
-type SonobuoyResultsWriter struct {
-	ResultsDir string
-	OutputFile string
-	Data       sono.Item
+// Done will tar up the results directory, write the done file which instructs Sonobuoy to
+// submit results to the aggregator.
+func Done() error{
+	dir := GetResultsDir()
+	outputFile := filepath.Join(dir, DefaultTarballName)
+	if err := tarball.DirToTarball(dir, outputFile,true ); err!=nil{
+		return fmt.Errorf("failed to tar up entire results directory: %w",err)
+	}
+	return WriteDone(outputFile)
 }
 
-func NewDefaultSonobuoyResultsWriter() SonobuoyResultsWriter {
-	return NewSonobuoyResultsWriter(os.Getenv("SONOBUOY_RESULTS_DIR"),defaultOutputFileName)
+func GetResultsDir() string{
+	return os.Getenv(SonobuoyResultsDirKey)
 }
 
-func NewSonobuoyResultsWriter(resultsDir, outputFile string) SonobuoyResultsWriter {
-	w := SonobuoyResultsWriter{
-		ResultsDir: resultsDir,
-		OutputFile: outputFile,
-		Data:       sono.Item{Items: []sono.Item{}},
+func WriteDone(resultsPath string) error{
+	if err:=os.WriteFile(filepath.Join(GetResultsDir(), DoneFileName), []byte(DefaultTarballName), 0666); err !=nil{
+		return fmt.Errorf("failed write done file: %w",err)
 	}
-	return w
-}
-
-func (w SonobuoyResultsWriter) AddTest(
-	testName string,
-	result string,
-	err error,
-	output string,
-) {
-	i := sono.Item{
-		Name:   testName,
-		Status: result,
-	}
-	if len(output) > 0 {
-		if i.Details == nil {
-			i.Details = map[string]interface{}{}
-		}
-		i.Details[sono.MetadataDetailsOutput] = output
-	}
-	if err != nil {
-		if i.Details == nil {
-			i.Details = map[string]interface{}{}
-		}
-		i.Details[sono.MetadataDetailsFailure] = err.Error()
-	}
-	w.Data.Items = append(w.Data.Items, i)
-}
-
-func (w SonobuoyResultsWriter) Done() error {
-	w.Data.Status=sono.AggregateStatus(w.Data.Items...)
-
-	outfile, err := os.Create(filepath.Join(w.ResultsDir, w.OutputFile))
-	if err != nil {
-		return errors.Wrap(err, "error creating results file")
-	}
-	defer outfile.Close()
-
-	enc := yaml.NewEncoder(outfile)
-	defer enc.Close()
-	err = enc.Encode(w.Data)
-	return errors.Wrap(err, "error writing to results file")
+	return nil
 }
